WIP not quite got rd_kafka_new yet getting too late
See the test for why it's not working.

Also tried with a variadic call signature, like

  rdkafkt_t * rd_kafka_new(rd_kafka_type_t type, ...)

but it won't let me put the errstr (which is in the ...)
in the OUTPUT section.

I want it to be where conf, errstr, and errstr_length are optional
and errstr gets an error message back. I'm not sure if this is even
the best way, though. Could also return a list like

  ($rk, $err) = rd_kafka_new($TYPE, ...);

but that seems annoying, too. I'd prefer to get $rk back directly
without having to do something else like dereference a hashref
or call a method on an object. I also don't like returning
two args because someone might call it in scalar context.
Looking at "want" for context, I don't like because people
end up putting things in hash assignments for example.

Maybe I'll just make it croak on error..